Remove `db` parameter (and other args) from the selector
Remove the `db` parameter (and accompanying args and kwargs) from
the interface selector decorator. Instead, the notion of a default
interface is created, such that a user simply generates a default
interface once for their app, and then this default interface is
always used implicitly. The one exception is for tests, in which
case the selector generates a new interface (including a new engine
pointing to a new temporary database) for each test.